*) reinserting type cast again
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@2623 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ source/de/anomic/htmlFilter/htmlFilterAbstractScraper.java
@@ -361,7 +361,7 @@ public abstract class htmlFilterAbstractScraper implements htmlFilterScraper {
             for  (int i = 0; i < bb.length(); i++) {
                 c = bb.charAt(i);
                 String z = code_iso8859s(c);
-                if (z == null) t.append(c); 
+                if (z == null) t.append((int)c); 
                 else t.append(z);
             }
             return t;
